+++ frontend/src/consts/index.js
@@ -1,5 +1,5 @@
 export const config = {
     SHOW_TEMPLATE: false,
-    // API_URL: "http://18.215.185.124/api"
-    API_URL: "http://localhost:8000/api"
+    API_URL: "http://18.215.185.124/api"
+    // API_URL: "http://localhost:8000/api"
 };
