+++ frontend/src/sections/@dashboard/files/index.js
@@ -1,0 +1,1 @@
+export { default as FilesListTemplates } from './FilesListTemplates';
